data: refactor crime data and geometry data
